chore: remove lazycat attempt
The use of lazy.nvim is far to nice to attempt to give it up just to
'nixify' my configuration. I don't need neovim to be fully self
contained when it comes to plugin configuration. I need it to be self
contained enough to have core language/binaries discoverable within the
path of neovim, ideally packages with this lazy.nvim configuration